+++ app/app.go
@@ -1258,6 +1258,7 @@ func (app *Evmos) setupUpgradeHandlers() {
 		v16.CreateUpgradeHandler(
 			app.mm, app.configurator,
 			app.EvmKeeper,
+			app.BankKeeper,
 			app.InflationKeeper,
 			app.GovKeeper,
 		),